Refactor and normalize rename op for drops
